+++ examples/info_core.rs
@@ -56,11 +56,11 @@ fn print_platform_device(platform: Platform, device: Device) {
 
     let mut event_list = EventList::new();
     kernel.cmd().enew(&mut event_list).enq().unwrap();
-    event_list.wait().unwrap();
+    event_list.wait_for().unwrap();
 
     let mut event = Event::empty();
     buffer.cmd().write(&vec![0.0; DIMS[0]]).enew(&mut event).enq().unwrap();
-    event.wait().unwrap();
+    event.wait_for().unwrap();
 
     println!("############### OpenCL Platform-Device Full Info ################");
     print!("\n");